+++ Elepay_StripeApplePay_Plugin.xcframework/ios-arm64_x86_64-simulator/Elepay_StripeApplePay_Plugin.framework/Headers/Elepay_StripeApplePay_Plugin.h
@@ -1,0 +1,18 @@
+//
+//  Elepay_StripeApplePay_Plugin.h
+//  Elepay-StripeApplePay-Plugin
+//
+//  Created by xuzhe on 2022/03/18.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for Elepay_StripeApplePay_Plugin.
+FOUNDATION_EXPORT double Elepay_StripeApplePay_PluginVersionNumber;
+
+//! Project version string for Elepay_StripeApplePay_Plugin.
+FOUNDATION_EXPORT const unsigned char Elepay_StripeApplePay_PluginVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Elepay_StripeApplePay_Plugin/PublicHeader.h>
+
+
