+++ Elepay_StripeApplePay_Plugin.xcframework/ios-arm64_x86_64-simulator/Elepay_StripeApplePay_Plugin.framework/Modules/module.modulemap
@@ -1,0 +1,11 @@
+framework module Elepay_StripeApplePay_Plugin {
+  umbrella header "Elepay_StripeApplePay_Plugin.h"
+  export *
+
+  module * { export * }
+}
+
+module Elepay_StripeApplePay_Plugin.Swift {
+  header "Elepay_StripeApplePay_Plugin-Swift.h"
+  requires objc
+}
