+++ Elepay_StripeApplePay_Plugin.xcframework/ios-arm64_x86_64-simulator/Elepay_StripeApplePay_Plugin.framework/_CodeSignature/CodeResources
@@ -1,0 +1,245 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/Elepay_StripeApplePay_Plugin-Swift.h</key>
+		<data>
+		D68IdB+T7/ypx7FmeRuHrfupGx4=
+		</data>
+		<key>Headers/Elepay_StripeApplePay_Plugin.h</key>
+		<data>
+		0mGufgitAPuDg5hvMgSp70Fc+dc=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		jEvmSQEJtrgrl4Lh5DTAnzSKq0w=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
+		<data>
+		2nAwUFT+WyzYoY4QmiXg9EmtG4c=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
+		<data>
+		uKTBgZzuLPF5hGvkbB6fyvC1+lU=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		dM2hLTqU4i28odimmPmboZotf+k=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		uKTBgZzuLPF5hGvkbB6fyvC1+lU=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		xSxPiavjIiRSA+bS0rXlRk/JJ9U=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
+		<data>
+		2nAwUFT+WyzYoY4QmiXg9EmtG4c=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
+		<data>
+		OjIZJWY6R/C1Z3OOWq5Ndlsww1o=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		zSUkiCOKqnKxK9mJZ4pL8lrVbls=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		OjIZJWY6R/C1Z3OOWq5Ndlsww1o=
+		</data>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		B8CkwbkaLXoqW1p/5YlvmIMRwlQ=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		+k7qlH9cceY+ibfMno4WCTAdGnA=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/Elepay_StripeApplePay_Plugin-Swift.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			BIgtlga334PD5a/+etjI0dNYJz0kE23IIwNDe3kYMPY=
+			</data>
+		</dict>
+		<key>Headers/Elepay_StripeApplePay_Plugin.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			Gpp2HugfY7EiQ8AmvqC5xen2WCY/XtyER/nbVFtU0ig=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			RPCXPsn0Y/rYwGhyoB1yVB2lP2BPVCPCePlhPSeFvbE=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			qopSE0UbQba2L0LjyrWiEDh5Z11HI3ZPLkbzmn/uXK4=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			dEbQCNL8PKKmHTwaBLeyxJZNgTeft3MUxG9BfvidILA=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			qopSE0UbQba2L0LjyrWiEDh5Z11HI3ZPLkbzmn/uXK4=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			gC6sQ0/U1HcwhzC5s79vsc5/TpPk7H2KC4PcG8xhSfg=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			RPCXPsn0Y/rYwGhyoB1yVB2lP2BPVCPCePlhPSeFvbE=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			YT4vXDwXA98m7qvA7+q9BfGRPoM5s5VzVZVThwSzN18=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			XLQichc+or1fmWMgemc83K7cw+Vn4kdHkZSwjSGbUkg=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			YT4vXDwXA98m7qvA7+q9BfGRPoM5s5VzVZVThwSzN18=
+			</data>
+		</dict>
+		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			7atvxMjIHwKLo5f09lRHxwkMz2xnklMkZ/fmqFI1KzQ=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			XUPLIv9W3xcmO4kzJJcQU/r/CSvPMvwbKCp5K7JWGeI=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
