--- Elepay_StripeApplePay_Plugin.xcframework/ios-x86_64-simulator/Elepay_StripeApplePay_Plugin.framework/_CodeSignature/CodeResources
@@ -1,201 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Headers/Elepay_StripeApplePay_Plugin-Swift.h</key>
-		<data>
-		uJ30jqjoHc/oLzjWKkuhr/XvGJ4=
-		</data>
-		<key>Headers/Elepay_StripeApplePay_Plugin.h</key>
-		<data>
-		0mGufgitAPuDg5hvMgSp70Fc+dc=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		Kf8minTTFsYblhj1RAyHFtmwwdo=
-		</data>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
-		<data>
-		5NyKG4j8WSIH7R1TeFexSAq9zCU=
-		</data>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
-		<data>
-		HquQNDQzXt2NKURZmtgQqsPwxuA=
-		</data>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
-		<data>
-		DLSvicuifu7Qrqwrtx5VpE8+Wy0=
-		</data>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64.swiftdoc</key>
-		<data>
-		5NyKG4j8WSIH7R1TeFexSAq9zCU=
-		</data>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64.swiftinterface</key>
-		<data>
-		HquQNDQzXt2NKURZmtgQqsPwxuA=
-		</data>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64.swiftmodule</key>
-		<data>
-		DLSvicuifu7Qrqwrtx5VpE8+Wy0=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		zNhfY7fy3lQqewz8+TcVQVj4N0E=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Headers/Elepay_StripeApplePay_Plugin-Swift.h</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			7H5q2CcnR4vm93rglgDDvoGjx4IrD7N3tUxqZomEXSU=
-			</data>
-		</dict>
-		<key>Headers/Elepay_StripeApplePay_Plugin.h</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			Gpp2HugfY7EiQ8AmvqC5xen2WCY/XtyER/nbVFtU0ig=
-			</data>
-		</dict>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			uc/YBRMpzKCeCaX9ypGV4fAHePM7bhaNeMO6HW+KQoM=
-			</data>
-		</dict>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			ZMTqYSLZDNYVmYTbzr2mfPNg7KZohS0xekGZh7xeodw=
-			</data>
-		</dict>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			PyguhdmRjzX4lO3ug+ggqQK8XssHAmTnlagVPZRmHyw=
-			</data>
-		</dict>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64.swiftdoc</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			uc/YBRMpzKCeCaX9ypGV4fAHePM7bhaNeMO6HW+KQoM=
-			</data>
-		</dict>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64.swiftinterface</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			ZMTqYSLZDNYVmYTbzr2mfPNg7KZohS0xekGZh7xeodw=
-			</data>
-		</dict>
-		<key>Modules/Elepay_StripeApplePay_Plugin.swiftmodule/x86_64.swiftmodule</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			PyguhdmRjzX4lO3ug+ggqQK8XssHAmTnlagVPZRmHyw=
-			</data>
-		</dict>
-		<key>Modules/module.modulemap</key>
-		<dict>
-			<key>hash2</key>
-			<data>
-			BeKQaJsiZoVX34WsyCk4DkOWep2SnOr08dpxMTUfiKM=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
