+++ Package.swift
@@ -4,7 +4,7 @@ import PackageDescription
 let package = Package(
     name: "Elepay_StripeApplePay_Plugin",
     platforms: [
-        .iOS(.v11)
+        .iOS(.v12)
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
